requirements.txt
asgiref==3.8.1
Django==4.2.20
django-formtools==2.5.1
sqlparse==0.5.3
typing_extensions==4.12.2
